Move database platforms out of Types namespace.
