+++ lib/CrEOF/Spatial/DBAL/Platform/AbstractPlatform.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (C) 2012 Derek J. Lambert
+ * Copyright (C) 2015 Derek J. Lambert
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,45 +21,39 @@
  * SOFTWARE.
  */
 
-namespace CrEOF\Spatial\DBAL\Types\Geography\Platforms;
+namespace CrEOF\Spatial\DBAL\Platform;
 
-use CrEOF\Spatial\DBAL\Types\Platforms\AbstractPlatform;
+use CrEOF\Geo\WKT\Parser as StringParser;
+use CrEOF\Geo\WKB\Parser as BinaryParser;
 use CrEOF\Spatial\Exception\InvalidValueException;
 use CrEOF\Spatial\PHP\Types\Geometry\GeometryInterface;
-use CrEOF\Spatial\PHP\Types\Geography\GeographyInterface;
 
 /**
- * PostgreSql spatial platform
+ * Abstract spatial platform
  *
  * @author  Derek J. Lambert <dlambert@dereklambert.com>
  * @license http://dlambert.mit-license.org MIT
  */
-class PostgreSql extends AbstractPlatform
+abstract class AbstractPlatform implements PlatformInterface
 {
-    const DEFAULT_SRID = 4326;
-
     /**
      * {@inheritdoc}
      */
-    public function getTypeFamily()
+    public function convertStringToPHPValue($sqlExpr)
     {
-        return GeographyInterface::GEOGRAPHY;
+        $parser = new StringParser($sqlExpr);
+
+        return $this->newObjectFromValue($parser->parse());
     }
 
     /**
      * {@inheritdoc}
      */
-    public function getSQLDeclaration(array $fieldDeclaration)
+    public function convertBinaryToPHPValue($sqlExpr)
     {
-        if ($fieldDeclaration['type']->getSQLType() == GeographyInterface::GEOGRAPHY) {
-            return 'geography';
-        }
+        $parser = new BinaryParser($sqlExpr);
 
-        if (isset($fieldDeclaration['srid'])) {
-            return sprintf('geography(%s,%d)', $fieldDeclaration['type']->getSQLType(), $fieldDeclaration['srid']);
-        }
-
-        return sprintf('geography(%s)', $fieldDeclaration['type']->getSQLType());
+        return $this->newObjectFromValue($parser->parse());
     }
 
     /**
@@ -67,35 +61,27 @@ class PostgreSql extends AbstractPlatform
      */
     public function convertToDatabaseValue(GeometryInterface $value)
     {
-        if ( ! ($value instanceof GeographyInterface)) {
-            throw InvalidValueException::invalidValueNotGeography();
-        }
-
-        if ($value->getSrid() === null) {
-            $value->setSrid(self::DEFAULT_SRID);
-        }
-
-        return sprintf(
-            'SRID=%d;%s(%s)',
-            $value->getSrid(),
-            strtoupper($value->getType()),
-            $value
-        );
+        return sprintf('%s(%s)', strtoupper($value->getType()), $value);
     }
 
     /**
-     * {@inheritdoc}
+     * Create spatial object from parsed value
+     *
+     * @param array $value
+     *
+     * @return GeometryInterface
+     * @throws \CrEOF\Spatial\Exception\InvalidValueException
      */
-    public function convertToPHPValueSQL($sqlExpr)
+    private function newObjectFromValue($value)
     {
-        return sprintf('ST_AsEWKT(%s)', $sqlExpr);
-    }
+        $constName = 'CrEOF\Spatial\PHP\Types\Geometry\GeometryInterface::' . strtoupper($value['type']);
 
-    /**
-     * {@inheritdoc}
-     */
-    public function convertToDatabaseValueSQL($sqlExpr)
-    {
-        return sprintf('ST_GeographyFromText(%s)', $sqlExpr);
+        if (! defined($constName)) {
+            throw InvalidValueException::unsupportedType($this->getTypeFamily(), strtoupper($value['type']));
+        }
+
+        $class = sprintf('CrEOF\Spatial\PHP\Types\%s\%s', $this->getTypeFamily(), constant($constName));
+
+        return new $class($value['value'], $value['srid']);
     }
 }
