+++ lib/CrEOF/Spatial/DBAL/Platform/PlatformInterface.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (C) 2012 Derek J. Lambert
+ * Copyright (C) 2015 Derek J. Lambert
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,36 +21,66 @@
  * SOFTWARE.
  */
 
-namespace CrEOF\Spatial\DBAL\Types\Geography\Platforms;
+namespace CrEOF\Spatial\DBAL\Platform;
 
-use CrEOF\Spatial\PHP\Types\Geography\GeographyInterface;
+use CrEOF\Spatial\PHP\Types\Geometry\GeometryInterface;
 
 /**
- * MySql spatial platform
+ * Spatial platform interface
  *
  * @author  Derek J. Lambert <dlambert@dereklambert.com>
  * @license http://dlambert.mit-license.org MIT
  */
-class MySql extends \CrEOF\Spatial\DBAL\Types\Geometry\Platforms\MySql
+interface PlatformInterface
 {
     /**
-     * {@inheritdoc}
+     * @param string $sqlExpr
+     *
+     * @return GeometryInterface
      */
-    public function getTypeFamily()
-    {
-        return GeographyInterface::GEOGRAPHY;
-    }
+    public function convertBinaryToPHPValue($sqlExpr);
 
     /**
-     * {@inheritdoc}
+     * @param string $sqlExpr
+     *
+     * @return GeometryInterface
      */
-    public function getSQLDeclaration(array $fieldDeclaration)
-    {
-        if ($fieldDeclaration['type']->getSQLType() == GeographyInterface::GEOGRAPHY) {
-            return 'GEOMETRY';
-        }
+    public function convertStringToPHPValue($sqlExpr);
 
-        return parent::getSQLDeclaration($fieldDeclaration);
-    }
+    /**
+     * @param GeometryInterface $value
+     *
+     * @return string
+     */
+    public function convertToDatabaseValue(GeometryInterface $value);
 
+    /**
+     * @param string $sqlExpr
+     *
+     * @return string
+     */
+    public function convertToDatabaseValueSQL($sqlExpr);
+
+    /**
+     * @param string $sqlExpr
+     *
+     * @return string
+     */
+    public function convertToPHPValueSQL($sqlExpr);
+
+    /**
+     * Get the type family for this interface (i.e. geometry or geography)
+     *
+     * @return string
+     */
+    public function getTypeFamily();
+
+    /**
+     * Gets the SQL declaration snippet for a field of this type.
+     *
+     * @param array $fieldDeclaration
+     *
+     * @return string
+     */
+    public function getSQLDeclaration(array $fieldDeclaration);
 }
